code to upload image to aws s3

+++ app/controllers/api_controller.rb
@@ -149,27 +149,40 @@ class ApiController < ApplicationController
           
         if user && user.authtoken_expiry > Time.now
           if user.photos.count < 3
-            rand_id = rand_string(20)
+            rand_id = rand_string(80)
             image_name = params[:image].original_filename
             image = MiniMagick::Image.read(params[:image].read)          
-          
-            img_public_path = "/public/uploads/#{rand_id}.png"
-            image_local_file_path = File.join(Rails.root, img_public_path)
-            image.write(image_local_file_path)
-          
-            photo = Photo.new(:name => image_name, :user_id => user.id, :title => params[:title], :image_url => "/uploads/#{rand_id}.png", :random_id => rand_id)
-          
-            if photo.save
-              render :json => photo.to_json
-            else
-              error_str = ""
-
-              @user.errors.each{|attr, msg|           
-                error_str += "#{attr} - #{msg},"
-              }
                     
-              e = Error.new(:status => 400, :message => error_str)
-              render :json => e.to_json, :status => 400
+            s3 = AWS::S3.new
+            
+            if s3
+              bucket = s3.buckets[ENV["S3_BUCKET_NAME"]]
+              
+              if !bucket
+                bucket = s3.buckets.create(ENV["S3_BUCKET_NAME"])
+              end
+              
+              s3_obj = bucket.objects[rand_id]
+              s3_obj.write(image, :acl => :public_read)
+              image_url = s3_obj.public_url
+                            
+              photo = Photo.new(:name => image_name, :user_id => user.id, :title => params[:title], :image_url => image_url, :random_id => rand_id)
+          
+              if photo.save
+                render :json => photo.to_json
+              else
+                error_str = ""
+
+                @user.errors.each{|attr, msg|           
+                  error_str += "#{attr} - #{msg},"
+                }
+                    
+                e = Error.new(:status => 400, :message => error_str)
+                render :json => e.to_json, :status => 400
+              end
+            else
+              e = Error.new(:status => 401, :message => "AWS S3 signature is wrong")
+              render :json => e.to_json, :status => 401              
             end
           else
             e = Error.new(:status => 403, :message => "You have already uploaded 3 photos!")
@@ -195,14 +208,21 @@ class ApiController < ApplicationController
           photo = Photo.where(:random_id => params[:photo_id]).first
           
           if photo && photo.user_id == user.id            
-            img_public_path = "/public#{photo.image_url}"
-            image_local_file_path = File.join(Rails.root, img_public_path)
+            s3 = AWS::S3.new
             
-            File.delete(image_local_file_path) if File.exist?(image_local_file_path)
-            photo.destroy
+            if s3
+              bucket = s3.buckets[ENV["S3_BUCKET_NAME"]]
+              s3_obj =  bucket.objects[photo.rand_id]
+              s3_obj.delete
+                            
+              photo.destroy
             
-            m = Message.new(:status => 200, :message => "Image is deleted.")          
-            render :json => m.to_json, :status => 200  
+              m = Message.new(:status => 200, :message => "Image is deleted.")          
+              render :json => m.to_json, :status => 200  
+            else
+              e = Error.new(:status => 401, :message => "AWS S3 signature is wrong")
+              render :json => e.to_json, :status => 401        
+            end                        
           else
             e = Error.new(:status => 401, :message => "You don't have permission to delete this photo!")
             render :json => e.to_json, :status => 401
